FBM Page Blurb plugin
This plugin helps to add meta title, meta excerpt and meta featured image to a page. The meta data can be added on the add/edit page view in the dashboard.

The shortcode [display-pages-blurb] displays the page blurb of the current page child pages.

The shortcode [display-pages-blurb id=\"xx\"] displays the page blurb of the page with id \'xx\'.